Decouple background layer to module.
