+++ public/js/loaders/level.js
@@ -1,6 +1,7 @@
 import {Matrix} from '../math.js';
 import Level from '../Level.js';
-import {createBackgroundLayer, createSpriteLayer} from '../layers.js';
+import {createBackgroundLayer} from '../layers/background.js';
+import {createSpriteLayer} from '../layers.js';
 import {loadJSON, loadSpriteSheet} from '../loaders.js';
 
 function setupCollision(levelSpec, level) {
